aws/lambda/oss_ci_aws_ce_tracking/requirements.txt
clickhouse_connect==0.8.5
boto3==1.35.33
python-dateutil==2.8.2
